requirements.txt
streamlit==1.28.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
plotly==5.17.0
pandas==2.0.3
openpyxl==3.1.2
scikit-learn==1.3.0
face-recognition==1.3.0